+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Changelog
 - Add `Concat` function for MySQL/PostgreSQL. (#873)
 - Patch for use_index/force_index mutable problem when making query. (#888)
 - Lift annotation field's priority in make query. (#883)
+- Make use/force index available in select type Query. (#893)
 
 0.17.6
 ------